@rtfeldman 2017-04-21
@evancz I'd love to see this get the same treatment as https://github.com/elm-lang/core/pull/850 😄