@evancz 2017-07-08
I read through this today and merged it into dev here https://github.com/elm-lang/core/commit/48e640e31617049e44ed712fe8097a406a0bb526

@mgold, I notice you use `(|>)` in a bunch of bitwise operations. Do you mind switching this code to not use that operator, instead just saying `Bitwise.and x y` like normal? It would be annoying to have a performance regression because we were depending on a couple optimizations happening in the right order, so I'd just feel more comfortable if we did not use anything extra.

That said, thanks again for your excellent work on this! Excited to see how it goes when we are testing out 0.19 :)