@mgold 2017-07-08
Pipe removal in #883.

Thanks so much for merging! You are quite welcome.